requirements.txt
streamlit==1.49.1
pandas==2.2.0
numpy==1.26.4
openpyxl==3.1.5
weasyprint==66.0
python-docx==1.2.0
Jinja2==3.1.6
Pillow==10.4.0
num2words==0.5.14
cairocffi==1.7.1
CairoSVG==2.8.2
tinycss2==1.4.0
cssselect2==0.8.0
python-dotenv==1.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
beautifulsoup4==4.12.3
lxml==5.3.0

# Contractor-specific dependencies
pdfplumber==0.11.0
pdf2image==1.17.0
pytesseract==0.3.10
opencv-python-headless==4.8.1.78

# AI-powered document input dependencies
google-cloud-vision==3.7.2
azure-cognitiveservices-vision-computervision==0.9.0
msrest==0.7.1
